+++ cli/azd/pkg/tools/azcli/ad.go
@@ -26,20 +26,6 @@ type AzureCredentials struct {
 	ResourceManagerEndpointUrl string `json:"resourceManagerEndpointUrl"`
 }
 
-func (cli *azCli) GetSignedInUserId(ctx context.Context) (*string, error) {
-	client, err := cli.createGraphClient(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	userProfile, err := client.Me().Get(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed retrieving current user profile: %w", err)
-	}
-
-	return &userProfile.Id, nil
-}
-
 func (cli *azCli) CreateOrUpdateServicePrincipal(
 	ctx context.Context,
 	subscriptionId string,
